+++ backend/src/main/java/com/backend/category/repository/CategoryRepository.java
@@ -1,6 +1,6 @@
-package com.backend.Category.repository;
+package com.backend.category.repository;
 
-import com.backend.Category.entity.Category;
+import com.backend.category.entity.Category;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface CategoryRepository extends JpaRepository<Category, Long> {
